Fix checkstyle violations and refactored unit tests
Unit tests which check that success of add command cannot be checked by
comparing with another addcommand instance with the same arguments as equality
of Expense is determined by a unique identifier randomly assigned.

Unit tests which used to check for failiure when editing an expense to have
the same fields as another should no longer fail as the identity of an Expense
is determined by its unique identifier which cannot be edited.