Add view tab to display resolve window. (#16)
* Add view tab to display resolve window.

Currently, the window just displays a text showing a success and
will block all inputs. The view tab is to be removed once the
product is finished

* Add relevant regions to resolve window.

Resolve window has the command box and a left, right panel.
Also includes some basic styling for the resolve window

* Add view tab to display resolve window.

Currently, the window just displays a text showing a success and
will block all inputs. The view tab is to be removed once the
product is finished

* Add relevant regions to resolve window.

Resolve window has the command box and a left, right panel.
Also includes some basic styling for the resolve window

* Update resolve window to display feedback.

Resolve window has a feedback box and cleared up a bug where if
the resolve window was closed and reopened, text from the previous
opened window is still present.

* Add css styling for text area in resolve window.

* Add documentation to resolve window and related parts

* Add EOF to FXML file

* Move getRoot to a parent variable.

* Change default window sizes

Add ExerHealth icon to resolve window and window will appear
with default sizes and centered on screen.

Closes #54