Changed how application stores commands for undo and redo
Application was storing commands for undo and redo in a Singleton class called VersionControl, which was updated in LogicManager after each invertible command execution, and was used to facilitate undo and redo commands in their execution. Undo and redo commands also failed silently when not being able to fully undo or redo their given number of iterations.

The use of a singleton introduced unecessary dependency and the silent failure of undo and redo commands were not as user friendly as it seems.

Lets change the way commands are stored with a HistoryManager that will be in ModelManager. History related methods for ModelManager to implement is defined in HistoryModel which is extended by Model. There is less dependency between LogicManager, UndoCommand, RedoCommand and ModelManager. Undo and Redo Commands are now atomic in nature, where they do not fail silently anymore and will not make partial changes, it will either be fully successful or fail completely.  "all" argument removed from undo and redo as it is not useful for user. History range is now set to 20 commands.