Change file hash to UUID (#68)
Previously we were using SHA-1 to hash the entire image when adding
images to weme. However, this means that you cannot have duplicated
images (which should not actually be our responsibility), and it makes
it more difficult to undo and redo newly added entries. By using UUID
instead, we can just stick to a file cleanup on exit to clear
unreferenced images, which is much cleaner than manually adding special
cases to each command for undo and redo.