Add feedback for Undoable commands
Rationale behind this is to inform users what command they just
undid / redid.

Implementation:
Add feedback in the form of strings to `versionedWeme`. The choice to
use Strings is to prevent coupling. While that makes it more useless for
now, its implementation can be extended to a `Command` or
`CommandResult` in the future if necessary.