Implement minimal version of Matching (#115)
* Create MatchCommandParser to parse MatchCommands

MatchCommands are meant to match patients to suitable donors
in the system.

* Add function to parse match commands

* Add skeleton of MatchCommand

* Add matching functionality to MatchCommand

This is an intermediate implementation of MatchCommand, which
is executed to match a patient with Donors in the Donor pool.
Currently when this command is executed, it displays all donors which
matches the specified patient.

Currently in this implementation, a match happens when patient and donor's
blood types are compatible and they pass the tissue typing test.

* Modify AddressBook.java to find and return Patient

* Retrieves specified Patient from PersonList

* Add blood type tests

* Add tissue typing tests

* Fix broken build

* Make edits based on PR Review

* Add unit test for MatchCommand

* Create a MatchCard to display Match Results

* Update PersonList with Match Results

* Render MatchCard to display matches

* Store results of matches

* Add NumberOfMatches to track matches of Patients

* Edit PersonListPanel to display matches

* Add MatchedDonor to represent matching donors

* Add MatchedPatient to represent number of matches

A MatchedPatient contains a Patient's details and the number
of matching donors with the Patient

* Edit Donor and Patient match result details

Previously, match results are stored in the Donor and Patient class.
With this commit, match results are no longer stored in
Donor and Patient classes, but are stored in MatchedDonor
and MatchedPatient classes.

* Change PersonListPanel for MatchCommands

Previously, matches and people were stored in the Logic's
filteredPersonsList. This resulted in bugs when rendering the
matches and people.

The solution is to create a separate list to store matches. The
MainWindow will select the list to display.

* Add a new list to store matches

* Add Matching functionality

The changes support matching individual Patients and all Patients

* Fixed failing build

* Add unit tests for MatchCommandParser.java

* Fix coding style

* Make change to messages

* Add additional tests to ModelManagerTest.java

* Add unit tests for MatchedPatient.java

* Make changes to MatchedPatient equals method

* Add unit test for MatchedDonor.java

* Edit equality check for MatchDonor

* Add unit tests for AddressBook and UniquePersonList

* Add tests for LogicManager

* Add integration tests for CommandResult and MatchCommand

* Modify test data for correct matching test results

* Modify ModelManager methods and CommandTestUtil