Fix bug where ui does not update
Listener in agenda listens to the list of days. Since day is a list
of activity, when the list of activity mutates, agenda fails to receive
an update because it only checks on days mutation and not the
internal mutation of day.

Temporarily fix the bug by forcing an update to
Itinerary when Ui focuses on itinerary