Add default case in switch on CommandContinuation.Type
This just exemplifies how useless Java's type system is.
There are only THREE POSSIBLE VALUES of the enum. This
switch is exhaustive. But you know, what if someone used
reflection to call the private constructor on
CommandContinuation.Type and then called the private constructor
on CommandContinuation just to screw with you, right?

Look, CS2103T coordinators: can we switch to a less useless
language? Please?