Cache meme cards to improve performance.
Previously, a MemeCard/TemplateCard/ImportMeme card is created anew
whenever the GridCell is updated. This results in a lot of lagging and
explosive memory usage because MemeCard/TemplateCard/ImportMeme reads
the image from the disk everytime it is created.

With this commit, MemeCard/TemplateCard/ImportMeme are cached in their
parents and no repeated disk IO is made. Memory usage now is about 500MB
for ~500 memes and does not increase when scrolling.

The respective GridViews can scroll smoothly now. Tested with 1000 memes
and no significant lag is observed.